Fix tx leak
